backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
aiohttp>=3.9.0
httpx>=0.27.0
pyjwt>=2.8.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.0
scrapling[fetchers]>=0.4.0
